[ive] hide IVE APIs by using PIMPL idom
Reviewed-on: https://10.240.0.84:8087/c/cviai/+/27159

+++ scripts/sdk_release.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
+set -e
+
 SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 CVIAI_ROOT=$(readlink -f $SCRIPT_DIR/../)
 TMP_WORKING_DIR=$CVIAI_ROOT/tmp
@@ -11,7 +13,7 @@ fi
 
 mkdir -p $TMP_WORKING_DIR/build_sdk
 pushd $TMP_WORKING_DIR/build_sdk
-wget ftp://swftp:cvitek@10.18.65.11/third_party/cmake/cmake-3.18.4-Linux-x86_64.tar.gz
+wget -c ftp://swftp:cvitek@10.18.65.11/third_party/cmake/cmake-3.18.4-Linux-x86_64.tar.gz
 tar zxf cmake-3.18.4-Linux-x86_64.tar.gz
 CMAKE_BIN=$PWD/cmake-3.18.4-Linux-x86_64/bin/cmake
 echo "Creating tmp working directory."
@@ -36,90 +38,49 @@ else
     exit 1
 fi
 
-# don't shrink opencv size if platform is 183x series
-if [[ "$CHIP_ARCH" == "CV182X" ]]; then
+
+if [[ "$CHIP_ARCH" == "CV183X" ]]; then
+    SHRINK_OPENCV_SIZE=OFF
+    USE_TPU_IVE=ON
+    KERNEL_ROOT=""
+elif [[ "$CHIP_ARCH" == "CV182X" ]]; then
     SHRINK_OPENCV_SIZE=ON
     OPENCV_INSTALL_PATH=""
-fi
-
-if [[ "$CHIP_ARCH" == "MARS" ]]; then
-    USE_IVE=OFF
-else
-    USE_IVE=ON
-fi
-
-if [[ "$CHIP_ARCH" == "MARS" ]]; then
+    USE_TPU_IVE=ON
+    KERNEL_ROOT=""
+    echo "CHIP_ARCH=CV182X"
+    echo "USE_TPU_IVE=${USE_TPU_IVE}"
+elif [[ "$CHIP_ARCH" == "MARS" ]]; then
+    SHRINK_OPENCV_SIZE=OFF # TODO: shrink opencv
+    USE_TPU_IVE=OFF
+    IVE_SDK_INSTALL_PATH=""
     if [[ "$SYSTEM_PROCESSOR" == "RISCV" ]]; then
         KERNEL_ROOT="${KERNEL_PATH}"/build/"${PROJECT_FULLNAME}"/riscv/usr/
-        $CMAKE_BIN -G Ninja $CVIAI_ROOT -DCVI_PLATFORM=$CHIP_ARCH \
-      	                                -DCVI_SYSTEM_PROCESSOR=$SYSTEM_PROCESSOR \
-                                        -DCMAKE_BUILD_TYPE=$BUILD_TYPE \
-                                        -DOPENCV_ROOT=$OPENCV_INSTALL_PATH \
-                                        -DENABLE_CVIAI_CV_UTILS=ON \
-                                        -DMLIR_SDK_ROOT=$TPU_SDK_INSTALL_PATH \
-                                        -DMIDDLEWARE_SDK_ROOT=$MW_PATH \
-                                        -DCMAKE_INSTALL_PREFIX=$AI_SDK_INSTALL_PATH \
-                                        -DTOOLCHAIN_ROOT_DIR=$HOST_TOOL_PATH \
-                                        -DCMAKE_TOOLCHAIN_FILE=$TOOLCHAIN_FILE \
-                                        -DSHRINK_OPENCV_SIZE=$SHRINK_OPENCV_SIZE \
-                                        -DKERNEL_ROOT=$KERNEL_ROOT \
-                                        -DUSE_IVE=$USE_IVE \
-                                        -DCVI_MIDDLEWARE_3RD_LDFLAGS="$CVI_TARGET_PACKAGES_LIBDIR" \
-                                        -DCVI_MIDDLEWARE_3RD_INCCLAGS="$CVI_TARGET_PACKAGES_INCLUDE"
-    elif [[ "$SYSTEM_PROCESSOR" == "ARM" ]]; then
+    else
         KERNEL_ROOT="${KERNEL_PATH}"/build/"${PROJECT_FULLNAME}"/arm/usr/
-        $CMAKE_BIN -G Ninja $CVIAI_ROOT -DCVI_PLATFORM=$CHIP_ARCH \
-      	                                -DCVI_SYSTEM_PROCESSOR=$SYSTEM_PROCESSOR \
-                                        -DCMAKE_BUILD_TYPE=$BUILD_TYPE \
-                                        -DOPENCV_ROOT=$OPENCV_INSTALL_PATH \
-                                        -DENABLE_CVIAI_CV_UTILS=ON \
-                                        -DMLIR_SDK_ROOT=$TPU_SDK_INSTALL_PATH \
-                                        -DMIDDLEWARE_SDK_ROOT=$MW_PATH \
-                                        -DCMAKE_INSTALL_PREFIX=$AI_SDK_INSTALL_PATH \
-                                        -DTOOLCHAIN_ROOT_DIR=$HOST_TOOL_PATH \
-                                        -DCMAKE_TOOLCHAIN_FILE=$TOOLCHAIN_FILE \
-                                        -DSHRINK_OPENCV_SIZE=$SHRINK_OPENCV_SIZE \
-                                        -DKERNEL_ROOT=$KERNEL_ROOT \
-                                        -DUSE_IVE=$USE_IVE \
-                                        -DCVI_MIDDLEWARE_3RD_LDFLAGS="$CVI_TARGET_PACKAGES_LIBDIR" \
-                                        -DCVI_MIDDLEWARE_3RD_INCCLAGS="$CVI_TARGET_PACKAGES_INCLUDE"
-
     fi
 else
-    if [[ "$USE_IVE" == "ON" ]]; then
-        $CMAKE_BIN -G Ninja $CVIAI_ROOT -DCVI_PLATFORM=$CHIP_ARCH \
-                                        -DCVI_SYSTEM_PROCESSOR=$SYSTEM_PROCESSOR \
-                                        -DCMAKE_BUILD_TYPE=$BUILD_TYPE \
-                                        -DOPENCV_ROOT=$OPENCV_INSTALL_PATH \
-                                        -DENABLE_CVIAI_CV_UTILS=ON \
-                                        -DMLIR_SDK_ROOT=$TPU_SDK_INSTALL_PATH \
-                                        -DMIDDLEWARE_SDK_ROOT=$MW_PATH \
-                                        -DIVE_SDK_ROOT=$IVE_SDK_INSTALL_PATH \
-                                        -DCMAKE_INSTALL_PREFIX=$AI_SDK_INSTALL_PATH \
-                                        -DTOOLCHAIN_ROOT_DIR=$HOST_TOOL_PATH \
-                                        -DCMAKE_TOOLCHAIN_FILE=$TOOLCHAIN_FILE \
-                                        -DSHRINK_OPENCV_SIZE=$SHRINK_OPENCV_SIZE \
-                                        -DUSE_IVE=$USE_IVE \
-                                        -DCVI_MIDDLEWARE_3RD_LDFLAGS="$CVI_TARGET_PACKAGES_LIBDIR" \
-                                        -DCVI_MIDDLEWARE_3RD_INCCLAGS="$CVI_TARGET_PACKAGES_INCLUDE"
-    else
-        $CMAKE_BIN -G Ninja $CVIAI_ROOT -DCVI_PLATFORM=$CHIP_ARCH \
-                                        -DCVI_SYSTEM_PROCESSOR=$SYSTEM_PROCESSOR \
-                                        -DCMAKE_BUILD_TYPE=$BUILD_TYPE \
-                                        -DOPENCV_ROOT=$OPENCV_INSTALL_PATH \
-                                        -DENABLE_CVIAI_CV_UTILS=ON \
-                                        -DMLIR_SDK_ROOT=$TPU_SDK_INSTALL_PATH \
-                                        -DMIDDLEWARE_SDK_ROOT=$MW_PATH \
-                                        -DCMAKE_INSTALL_PREFIX=$AI_SDK_INSTALL_PATH \
-                                        -DTOOLCHAIN_ROOT_DIR=$HOST_TOOL_PATH \
-                                        -DCMAKE_TOOLCHAIN_FILE=$TOOLCHAIN_FILE \
-                                        -DSHRINK_OPENCV_SIZE=$SHRINK_OPENCV_SIZE \
-                                        -DUSE_IVE=$USE_IVE \
-                                        -DCVI_MIDDLEWARE_3RD_LDFLAGS="$CVI_TARGET_PACKAGES_LIBDIR" \
-                                        -DCVI_MIDDLEWARE_3RD_INCCLAGS="$CVI_TARGET_PACKAGES_INCLUDE"
-    fi
+    echo "Unsupported chip architecture: ${CHIP_ARCH}"
+fi
 
-fi   
+$CMAKE_BIN -G Ninja $CVIAI_ROOT -DCVI_PLATFORM=$CHIP_ARCH \
+                                        -DCVI_SYSTEM_PROCESSOR=$SYSTEM_PROCESSOR \
+                                        -DCMAKE_BUILD_TYPE=$BUILD_TYPE \
+                                        -DOPENCV_ROOT=$OPENCV_INSTALL_PATH \
+                                        -DENABLE_CVIAI_CV_UTILS=ON \
+                                        -DMLIR_SDK_ROOT=$TPU_SDK_INSTALL_PATH \
+                                        -DMIDDLEWARE_SDK_ROOT=$MW_PATH \
+                                        -DTPU_IVE_SDK_ROOT=$IVE_SDK_INSTALL_PATH \
+                                        -DCMAKE_INSTALL_PREFIX=$AI_SDK_INSTALL_PATH \
+                                        -DTOOLCHAIN_ROOT_DIR=$HOST_TOOL_PATH \
+                                        -DCMAKE_TOOLCHAIN_FILE=$TOOLCHAIN_FILE \
+                                        -DSHRINK_OPENCV_SIZE=$SHRINK_OPENCV_SIZE \
+                                        -DKERNEL_ROOT=$KERNEL_ROOT \
+                                        -DUSE_TPU_IVE=$USE_TPU_IVE \
+                                        -DCVI_MIDDLEWARE_3RD_LDFLAGS="$CVI_TARGET_PACKAGES_LIBDIR" \
+                                        -DCVI_MIDDLEWARE_3RD_INCCLAGS="$CVI_TARGET_PACKAGES_INCLUDE"
+
+
 ninja -j8 || exit 1
 ninja install || exit 1
 popd
@@ -139,17 +100,20 @@ for v in $CVI_TARGET_PACKAGES_LIBDIR; do
 done
 
 pushd ${AI_SDK_INSTALL_PATH}/module/app
-make MW_PATH="$MW_PATH" TPU_PATH="$TPU_SDK_INSTALL_PATH" IVE_PATH="$IVE_SDK_INSTALL_PATH" USE_IVE=$USE_IVE SYSTEM_PROCESSOR=$SYSTEM_PROCESSOR CHIP=$CHIP_ARCH -j10 || exit 1
+make MW_PATH="$MW_PATH" TPU_PATH="$TPU_SDK_INSTALL_PATH" IVE_PATH="$IVE_SDK_INSTALL_PATH" USE_TPU_IVE="$USE_TPU_IVE" SYSTEM_PROCESSOR=$SYSTEM_PROCESSOR CHIP=$CHIP_ARCH -j10 || exit 1
 make install || exit 1
 make clean || exit 1
 echo "done"
 popd
 
-pushd ${AI_SDK_INSTALL_PATH}/sample
-make MW_PATH="$MW_PATH" TPU_PATH="$TPU_SDK_INSTALL_PATH" IVE_PATH="$IVE_SDK_INSTALL_PATH" CVI_TRACER_PATH="$CVI_TRACER_ROOT_PATH" USE_IVE=$USE_IVE CHIP=$CHIP_ARCH SDK_VER=$SDK_VER SYSTEM_PROCESSOR=$SYSTEM_PROCESSOR USE_IVE=$USE_IVE -j10 || exit 1
-make install || exit 1
-make clean || exit 1
-echo "done"
-popd
+# skip building samples with Makefile temporarily, because cannot find correct linker here.
+if [[ "$SYSTEM_PROCESSOR" != "RISCV" ]]; then
+    pushd ${AI_SDK_INSTALL_PATH}/sample
+    make MW_PATH="$MW_PATH" TPU_PATH="$TPU_SDK_INSTALL_PATH" IVE_PATH="$IVE_SDK_INSTALL_PATH" CVI_TRACER_PATH="$CVI_TRACER_ROOT_PATH" USE_TPU_IVE=$USE_TPU_IVE CHIP=$CHIP_ARCH SDK_VER=$SDK_VER SYSTEM_PROCESSOR=$SYSTEM_PROCESSOR -j10 || exit 1
+    make install || exit 1
+    make clean || exit 1
+    echo "done"
+    popd
+fi
 
 rm -rf ${AI_SDK_INSTALL_PATH}/tmp_install
